Check if Calico `canReachHost` parameter is valid
The `canReachHost` represents a string representation of an IP. In the
current implementation it is not guaranteed that the passed parameter is
actually a valid IP address and would automatically classify it as IPv6.

The function will now check if the IP is valid and return an appropiate
error message if not.